chore(scripts): move record to scripts
