+++ pyproject.toml
@@ -165,7 +165,7 @@ all = [
 lerobot-calibrate="lerobot.calibrate:main"
 lerobot-find-cameras="lerobot.scripts.lerobot_find_cameras:main"
 lerobot-find-port="lerobot.scripts.lerobot_find_port:main"
-lerobot-record="lerobot.record:main"
+lerobot-record="lerobot.scripts.lerobot_record:main"
 lerobot-replay="lerobot.replay:main"
 lerobot-setup-motors="lerobot.setup_motors:main"
 lerobot-teleoperate="lerobot.teleoperate:main"
