+++ tests/test_control_robot.py
@@ -17,8 +17,8 @@
 from unittest.mock import patch
 
 from lerobot.calibrate import CalibrateConfig, calibrate
-from lerobot.record import DatasetRecordConfig, RecordConfig, record
 from lerobot.replay import DatasetReplayConfig, ReplayConfig, replay
+from lerobot.scripts.lerobot_record import DatasetRecordConfig, RecordConfig, record
 from lerobot.teleoperate import TeleoperateConfig, teleoperate
 from tests.fixtures.constants import DUMMY_REPO_ID
 from tests.mocks.mock_robot import MockRobotConfig
